Fixed a crash in `keyboarddisplay` module when notes greater than 127 in value where drawn. Tweaked the display so it can't render notes with a value greater than 127. Fixed the octave note text being off by 2 octaves.

+++ Source/KeyboardDisplay.h
@@ -94,7 +94,7 @@ private:
    bool mShowScale{ false };
    bool mGetVelocityFromClickHeight{ false };
    bool mHideLabels{ false };
-   std::array<float, 128> mLastOnTime{};
-   std::array<float, 128> mLastOffTime{};
+   std::array<double, 128> mLastOnTime{};
+   std::array<double, 128> mLastOffTime{};
    std::unordered_map<int, int> mKeyPressRegister{};
 };
